requirements.txt
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-spreadsheet==0.0.6
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
protobuf==4.21.7
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.5.1
tornado==6.2
traitlets==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
watchdog==2.1.9
wcwidth==0.2.5